Ability to benchmark vanilla executions
The dmmv.cu has a commented part that calculates time spent by GPU just doing GEMV. When you uncomment this, it prints out the value on screen. When commented it doesnt.

Thing is that it comments after every GEMV execution which is not good. So for now, I have kept things this way.

Need to come up with a better way to do this.

We have the Timer Class in the record replay mode branch, we could port those changes over here.